requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
altair==4.1.0
apex==0.1
appdirs==1.4.4
asgiref==3.2.7
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==19.3.0
audioread==3.0.0
av==10.0.0
backcall==0.1.0
base58==2.0.0
bcolors==1.0.4
bert4keras==0.11.5
bleach==3.1.4
blessed==1.19.1
blinker==1.4
brotlipy==0.7.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cmake==3.16.3
colorama==0.4.3
coloredlogs==15.0.1
cryptography==3.4.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
dlib==19.24.0
docopt==0.6.2
docutils==0.15.2
einops==0.6.1
entrypoints==0.3
enum-compat==0.0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.2.2
Flask-Cors==4.0.1
flatbuffers==23.5.26
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.51.0
gpustat==1.0.0
grpcio==1.50.0
h5py==3.7.0
httplib2==0.17.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==2.9
imageio==2.21.3
imgaug==0.4.0
importlib-metadata==6.7.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.1.2
jedi==0.16.0
Jinja2==3.1.2
jmespath==0.9.5
joblib==1.2.0
jpegio==0.2.4
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
kmeans-pytorch==0.3
kornia==0.6.12
libclang==16.0.6
librosa==0.9.2
lightning-utilities==0.10.1
llvmlite==0.38.0
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mdurl==0.1.2
memory-profiler==0.61.0
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv-lite==2.0.0
mmengine==0.7.4
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
natsort==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.6.3
notebook==6.0.3
numba==0.48.0
numpy==1.21.6
nvidia-ml-py==11.495.46
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.14.1
opencv-python==4.1.2.30
opencv-python-headless==4.6.0.66
openmim==0.3.7
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==3.8.0
pooch==1.6.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.20.0
psutil==5.9.2
ptflops==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.21
pydeck==0.3.0b3
pydub==0.25.1
Pygments==2.15.1
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.11
pytorch-lightning==1.9.5
pytorchtools==0.0.2
pytz==2019.3
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==19.0.0
qudida==0.0.4
RandAugment==0.1
regex==2023.8.8
requests==2.23.0
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.4.2
rsa==4.0
s3transfer==0.3.3
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segment-anything==1.0
Send2Trash==1.5.0
Shapely==1.8.4
six==1.14.0
sklearn==0.0
soundfile==0.11.0
soupsieve==2.0
sqlparse==0.3.1
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.3.2
tokenizers==0.13.3
toml==0.10.0
tomli==2.0.1
toolz==0.10.0
torch==1.12.1+cu116
torch-dct==0.1.6
torch-kmeans==0.2.0
torchaudio==0.12.1+cu116
torchinfo==1.7.1
torchmetrics==0.11.4
torchvision==0.13.1+cu116
tornado==5.1.1
tqdm==4.66.1
traitlets==4.3.3
transformers==4.30.2
typing_extensions==4.6.3
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
validators==0.14.2
watchdog==0.10.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.5.1
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.1
yarg==0.1.9
yarl==1.9.4
zipp==3.9.0
